+++ src/main/java/com/swiggy/app/Service/ItemServiceImpl.java
@@ -2,6 +2,7 @@ package com.swiggy.app.Service;
 
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Objects;
 
 import org.springframework.stereotype.Service;
 
@@ -35,18 +36,29 @@ private ItemRepository itemRepo;
 	
 	@Override
 	public Item updateItem(Item item,Long id) {
-		Boolean exists = itemRepo.existsById(id);
-		if (exists) {
-			return itemRepo.save(item);
-		} 
-		return null;
+		Item item1 = itemRepo.findById(id).get();
+		if (item1 != null) {
+			if (item.getName() != null) {
+				item1.setName(item.getName());
+			}
+			if (!Objects.isNull(item.getPrice())) {
+				item1.setPrice(item.getPrice());
+			}
+			return item1;
+		} else {
+			itemRepo.save(item);
+			return item;
+		}
 	}
 	
 	@Override
-	public Item deleteItem(Long id) {
-		itemRepo.findById(id).orElseThrow(()->new NoSuchElementException());
-		Item rest = itemRepo.findById(id).get();
-		 itemRepo.deleteById(id);
-		 return rest;
+	public String deleteItem(Long id) {
+		if(id!=null) {
+			itemRepo.deleteById(id);
+			return "deleted successfully!";
+		}
+		else {
+			throw new NoSuchElementException();
+		}
 	}
 }
